Adição da classe Vector2

+++ include/Player.hpp
@@ -7,7 +7,7 @@ class Player : public GameObject{
 	
 public:
 
-	void load(int x_pos, int y_pos, int width, int height, std::string Id_Texture);
+	void load(float x_pos, float y_pos, int width, int height, std::string Id_Texture);
 
 	void draw(SDL_Renderer* m_Renderer);
